- Create function for writing classification report. - Plot heatmaps generated during writing of the report.
